Change redirection to default view and put default view to protected routes
