+++ frontend/src/components/Login.jsx
@@ -3,6 +3,7 @@ import { Link, useNavigate } from 'react-router-dom'
 import { API } from '../api'
 import { getCookie } from '../helper'
 import { useAuth } from '../contexts/AuthContext'
+import ConnectBackend from './ConnectBackend'
 
 const Login = () => {
   const [formData, setFormData] = useState({
@@ -30,7 +31,7 @@ const Login = () => {
 
     if (response.status === 200) {
       setLoading(true)
-      navigate('/view')
+      navigate('/')
     }
     else {
       const data = await response.json()
@@ -40,6 +41,7 @@ const Login = () => {
 
   return (
     <div>
+      <ConnectBackend />
       <h1>Login your account</h1>
       <form onSubmit={login} className='form'>
         <div className='error'>{error}</div>
