+++ frontend/src/components/NonProtectedRoute.jsx
@@ -5,7 +5,7 @@ const NonProtectedRoute = () => {
   const { user, loading } = useAuth()
 
   if (loading) { return <div>Loading...!</div> }
-  if (user) { return <Navigate to='/view' /> }
+  if (user) { return <Navigate to='/' /> }
 
   return <Outlet />
 }
